Refactor: Standardize parsed item initialization in Shopify creators
This commit refactors the initialization of `Shopify::SaleCreator` and `Shopify::ProductCreator` to accept `parsed_item:` instead of `parsed_order:` and `parsed_product:` respectively. This change aligns the initialization parameters with the `api_item` parameter used in the parsers and `BasePullJob`, promoting consistency across the Shopify integration.

Additionally, the commit updates `Shopify::BasePullJob` to pass `parsed_item:` to the creator classes. It also simplifies the iteration variable name in `Shopify::PullEditionsJob` from `variant` to `pe` for brevity.