Feat: Add Audit Notifications to Models
This commit adds the `HasAuditNotifications` concern to several models to enable sending notifications to Slack when audited records are created, updated, or destroyed.

The following models were updated to include the `HasAuditNotifications` concern:

- Color
- Version
- Shape
- Payment
- Warehouse
- Notification
- Size
- ProductColor
- ProductBrand
- ProductSupplier
- ProductVersion
- ShippingCompany
- ProductSize
- ProductSale
- Supplier
- Edition
- PurchasedProduct
- Franchise
- Purchase
- WarehouseTransition
- Config
- Brand
- Customer
- Product
- Sale

This change ensures that audit logs for these models will trigger notifications, providing better visibility into data changes within the application.

Additionally, the `Shop` model was updated to inherit from `ApplicationRecord` instead of `ActiveRecord::Base`.

The `woo_id_is_valid?` methods in the `Customer` model were reordered.

The `Brand` model now includes the `Sanitizable` concern before `HasAuditNotifications`.

The `enum` definition in the `Config` model was updated to use explicit values.

The `update_full_title` method in the `Product` model was moved below the `generate_full_title` method.

The `prev_image_id` method in the `Product` model was updated to use the `...` operator for id comparison.

The methods in the `Sale` model were reordered.